ClothingMod Transmogger compatibility patch
This commit changes the way items are named internally.

Using the lua script, the items will still be renamed, in order to work properly with all localizations.

Internally the item names are stored in the english localization natively, so that the transmogger can use the items added by the ClothingMod.

In order to achieve this I have added code to handle the localization modification and serialization, as well as changing simtype IDs, so that they are stored sorted in the resource descriptor files.

While sorting the IDs is not strictly required for the game in all cases, they are sorted by default and the SDK the Transmogger uses needs them to be sorted.